over writed

+++ ACH_CreateNifti.m
@@ -1,4 +1,4 @@
-function ACH_CreateNifti(SUBJECT)
+function ACH_CreateNifti
 
 % Transforming dicom files to nii.gz file obtained by SIEMENS 3T scanner in
 % Tamagawa University, Machida, JAPAN
@@ -15,25 +15,27 @@ function ACH_CreateNifti(SUBJECT)
 
 %% change file format 
 
-if notDefined('SUBJECT')
+% if notDefined('SUBJECT')
  [~, subDir] = fileparts(pwd);
  SUBJECT = fullfile('/home/ganka/dMRI_data/',subDir);
  T1 = fullfile(SUBJECT,'T1');
  dwi1st = fullfile(SUBJECT,'dwi1st');
  dwi2nd = fullfile(SUBJECT,'dwi2nd');
- niiFolder = pwd;
-end
+
+% end
 
 outFormat = 'nii.gz';
 
 % create nii.gz files 
-dicm2nii(T1, niiFolder, outFormat)
+if ~exist('t1.nii.gz')
+ dicm2nii(T1, SUBJECT, outFormat)
+end
 dicm2nii(dwi1st, dwi1st, outFormat)
 dicm2nii(dwi2nd, dwi2nd, outFormat)
 
 %% make raw foldef under SUBJECT folder
 
-RawFolder = fullfile(pwd,'raw');
+RawFolder = fullfile(SUBJECT,'raw');
 if ~exist(RawFolder)
  mkdir(fullfile(SUBJECT,'raw'))
 end
@@ -47,7 +49,7 @@ Dwi1st = dir(fullfile(dwi1st,'*iso.nii.gz'));
 copyfile(fullfile(dwi1st,Bvec.name),fullfile(RawFolder,'dwi1st.bvec'));
 copyfile(fullfile(dwi1st,Bval.name),fullfile(RawFolder,'dwi1st.bval'));
 copyfile(fullfile(dwi1st,Dwi1st.name),fullfile(RawFolder,'dwi1st.nii.gz'));
-
+clear Bval Bvec
 % dwi2nd
 Bvec = dir(fullfile(dwi2nd,'*.bvec')); 
 Bval = dir(fullfile(dwi2nd,'*.bval'));
@@ -57,8 +59,9 @@ copyfile(fullfile(dwi2nd,Bval.name),fullfile(RawFolder,'dwi2nd.bval'));
 copyfile(fullfile(dwi2nd,Dwi2nd.name),fullfile(RawFolder,'dwi2nd.nii.gz'));
 
 % T1
-T1 = dir(fullfile(SUBJECT,'t1*iso.nii.gz'));
-copyfile(T1.name,'t1.nii.gz');
-
+if ~exist('t1.nii.gz')
+ T1 = dir(fullfile(SUBJECT,'t1*iso.nii.gz'));
+ copyfile(T1.name,'t1.nii.gz');
+end
 % See also ACH_Preprocess
